[Feat] #8 - CompositionalFactory 추가
